chore: refactor services and dependency containers to support multiple verticals (#795)
