+++ apps/api/routes/crm/v1/event.ts
@@ -22,7 +22,9 @@ import { ListParams } from '@supaglue/types';
 import { camelcaseKeysSansCustomFields } from '@supaglue/utils/camelcase';
 import { Request, Response, Router } from 'express';
 
-const { eventService } = getDependencyContainer();
+const {
+  crm: { eventService },
+} = getDependencyContainer();
 
 export default function init(app: Router): void {
   const router = Router();
