+++ apps/sync-worker/dependency_container.ts
@@ -1,16 +1,14 @@
 import { getCoreDependencyContainer } from '@supaglue/core';
+import { ConnectionService, IntegrationService, RemoteService, SyncHistoryService } from '@supaglue/core/services';
 import {
   AccountService,
-  ConnectionService,
-  ContactService,
+  ContactService as CrmContactService,
   EventService,
-  IntegrationService,
   LeadService,
   OpportunityService,
-  RemoteService,
-  SyncHistoryService,
   UserService,
-} from '@supaglue/core/services';
+} from '@supaglue/core/services/common_models/crm';
+import { ContactService as EngagementContactService } from '@supaglue/core/services/common_models/engagement';
 import type { PrismaClient } from '@supaglue/db';
 import { Client, Connection } from '@temporalio/client';
 import fs from 'fs';
@@ -20,36 +18,30 @@ type DependencyContainer = {
   prisma: PrismaClient;
   temporalClient: Client;
   connectionService: ConnectionService;
-  contactService: ContactService;
   remoteService: RemoteService;
-  accountService: AccountService;
-  leadService: LeadService;
-  userService: UserService;
-  eventService: EventService;
-  opportunityService: OpportunityService;
   syncService: SyncService;
   syncHistoryService: SyncHistoryService;
   integrationService: IntegrationService;
   applicationService: ApplicationService;
+  crm: {
+    contactService: CrmContactService;
+    accountService: AccountService;
+    leadService: LeadService;
+    userService: UserService;
+    eventService: EventService;
+    opportunityService: OpportunityService;
+  };
+  engagement: {
+    contactService: EngagementContactService;
+  };
 };
 
 // global
 let dependencyContainer: DependencyContainer | undefined = undefined;
 
 function createDependencyContainer(): DependencyContainer {
-  const {
-    prisma,
-    connectionService,
-    contactService,
-    remoteService,
-    accountService,
-    leadService,
-    userService,
-    eventService,
-    opportunityService,
-    syncHistoryService,
-    integrationService,
-  } = getCoreDependencyContainer();
+  const { prisma, connectionService, remoteService, syncHistoryService, integrationService, crm, engagement } =
+    getCoreDependencyContainer();
 
   const TEMPORAL_ADDRESS =
     process.env.SUPAGLUE_TEMPORAL_HOST && process.env.SUPAGLUE_TEMPORAL_PORT
@@ -81,16 +73,12 @@ function createDependencyContainer(): DependencyContainer {
     temporalClient,
     applicationService,
     connectionService,
-    contactService,
     remoteService,
-    accountService,
-    leadService,
-    userService,
-    eventService,
-    opportunityService,
     syncService,
     syncHistoryService,
     integrationService,
+    crm,
+    engagement,
   };
 }
