+++ packages/core/dependency_container.ts
@@ -2,19 +2,22 @@ import prisma, { PrismaClient } from '@supaglue/db';
 import fs from 'fs';
 import { Pool } from 'pg';
 import {
-  AccountService,
   ConnectionService,
-  ContactService,
   CustomerService,
-  EventService,
   IntegrationService,
-  LeadService,
-  OpportunityService,
   RemoteService,
   SgUserService,
   SyncHistoryService,
-  UserService,
 } from './services';
+import {
+  AccountService,
+  ContactService as CrmContactService,
+  EventService,
+  LeadService,
+  OpportunityService,
+  UserService,
+} from './services/common_models/crm';
+import { ContactService as EngagementContactService } from './services/common_models/engagement';
 
 export type CoreDependencyContainer = {
   pgPool: Pool;
@@ -26,15 +29,22 @@ export type CoreDependencyContainer = {
   integrationService: IntegrationService;
   customerService: CustomerService;
   remoteService: RemoteService;
+  syncHistoryService: SyncHistoryService;
 
   // crm
-  accountService: AccountService;
-  contactService: ContactService;
-  leadService: LeadService;
-  opportunityService: OpportunityService;
-  userService: UserService;
-  eventService: EventService;
-  syncHistoryService: SyncHistoryService;
+  crm: {
+    accountService: AccountService;
+    contactService: CrmContactService;
+    leadService: LeadService;
+    opportunityService: OpportunityService;
+    userService: UserService;
+    eventService: EventService;
+  };
+
+  // engagement
+  engagement: {
+    contactService: EngagementContactService;
+  };
 };
 
 // global
@@ -76,10 +86,13 @@ function createCoreDependencyContainer(): CoreDependencyContainer {
   const accountService = new AccountService(pgPool, prisma, remoteService);
   const leadService = new LeadService(pgPool, prisma, remoteService);
   const opportunityService = new OpportunityService(pgPool, prisma, remoteService);
-  const contactService = new ContactService(pgPool, prisma, remoteService);
+  const contactService = new CrmContactService(pgPool, prisma, remoteService);
   const syncHistoryService = new SyncHistoryService(prisma, connectionService);
   const userService = new UserService(pgPool, prisma, remoteService);
   const eventService = new EventService(pgPool, prisma, remoteService);
+
+  // engagement
+  const engagementContactService = new EngagementContactService(pgPool, prisma, remoteService);
 
   return {
     pgPool,
@@ -90,14 +103,20 @@ function createCoreDependencyContainer(): CoreDependencyContainer {
     customerService,
     integrationService,
     remoteService,
-    // crm
-    contactService,
-    accountService,
-    leadService,
-    opportunityService,
-    userService,
-    eventService,
     syncHistoryService,
+    // crm
+    crm: {
+      contactService,
+      accountService,
+      leadService,
+      opportunityService,
+      userService,
+      eventService,
+    },
+    // engagement
+    engagement: {
+      contactService: engagementContactService,
+    },
   };
 }
