+++ packages/core/services/common_models/crm/index.ts
@@ -1,0 +1,6 @@
+export * from './account_service';
+export * from './contact_service';
+export * from './event_service';
+export * from './lead_service';
+export * from './opportunity_service';
+export * from './user_service';
