+++ packages/core/services/common_models/crm/lead_service.ts
@@ -1,48 +1,47 @@
 import { COMMON_MODEL_DB_TABLES } from '@supaglue/db';
 import type {
-  Contact,
-  ContactCreateParams,
-  ContactFilters,
-  ContactUpdateParams,
   GetInternalParams,
+  Lead,
+  LeadCreateParams,
+  LeadFilters,
+  LeadUpdateParams,
   ListInternalParams,
   PaginatedResult,
   SearchInternalParams,
 } from '@supaglue/types';
 import { Readable } from 'stream';
 import { v5 as uuidv5 } from 'uuid';
-import { NotFoundError, UnauthorizedError } from '../../errors';
-import { getPaginationParams, getPaginationResult } from '../../lib/pagination';
-import { getRemoteId } from '../../lib/remote_id';
-import { fromContactModel, fromRemoteContactToDbContactParams } from '../../mappers';
-import { CrmRemoteClient } from '../../remotes/crm/base';
-import { CommonModelBaseService, getLastModifiedAt, UpsertRemoteCommonModelsResult } from './base_service';
+import { CommonModelBaseService, getLastModifiedAt, UpsertRemoteCommonModelsResult } from '..';
+import { NotFoundError, UnauthorizedError } from '../../../errors';
+import { getPaginationParams, getPaginationResult, getRemoteId } from '../../../lib';
+import { fromLeadModel, fromRemoteLeadToDbLeadParams } from '../../../mappers';
+import { CrmRemoteClient } from '../../../remotes/crm/base';
 
-export class ContactService extends CommonModelBaseService {
+export class LeadService extends CommonModelBaseService {
   public constructor(...args: ConstructorParameters<typeof CommonModelBaseService>) {
     super(...args);
   }
 
-  public async getById(id: string, connectionId: string, getParams: GetInternalParams): Promise<Contact> {
-    const model = await this.prisma.crmContact.findUnique({
+  public async getById(id: string, connectionId: string, getParams: GetInternalParams): Promise<Lead> {
+    const model = await this.prisma.crmLead.findUnique({
       where: { id },
     });
     if (!model) {
-      throw new NotFoundError(`Can't find contact with id: ${id}`);
+      throw new NotFoundError(`Can't find Lead with id: ${id}`);
     }
     if (model.connectionId !== connectionId) {
       throw new UnauthorizedError('Unauthorized');
     }
-    return fromContactModel(model, getParams);
+    return fromLeadModel(model, getParams);
   }
 
   public async search(
     connectionId: string,
     searchParams: SearchInternalParams,
-    filters: ContactFilters
-  ): Promise<PaginatedResult<Contact>> {
+    filters: LeadFilters
+  ): Promise<PaginatedResult<Lead>> {
     const { page_size, cursor } = searchParams;
-    const models = await this.prisma.crmContact.findMany({
+    const models = await this.prisma.crmLead.findMany({
       ...getPaginationParams(page_size, cursor),
       where: {
         connectionId,
@@ -57,18 +56,18 @@ export class ContactService extends CommonModelBaseService {
         id: 'asc',
       },
     });
-    const results = models.map((model) => fromContactModel(model, searchParams));
+    const results = models.map((model) => fromLeadModel(model, searchParams));
     return {
       ...getPaginationResult(page_size, cursor, results),
       results,
     };
   }
 
-  public async list(connectionId: string, listParams: ListInternalParams): Promise<PaginatedResult<Contact>> {
+  // TODO: implement rest of list params
+  public async list(connectionId: string, listParams: ListInternalParams): Promise<PaginatedResult<Lead>> {
     const { page_size, cursor, include_deleted_data, created_after, created_before, modified_after, modified_before } =
       listParams;
-
-    const models = await this.prisma.crmContact.findMany({
+    const models = await this.prisma.crmLead.findMany({
       ...getPaginationParams(page_size, cursor),
       where: {
         connectionId,
@@ -86,107 +85,103 @@ export class ContactService extends CommonModelBaseService {
         id: 'asc',
       },
     });
-    const results = models.map((model) => fromContactModel(model, listParams));
+    const results = models.map((model) => fromLeadModel(model, listParams));
     return {
       ...getPaginationResult(page_size, cursor, results),
       results,
     };
   }
 
-  public async create(customerId: string, connectionId: string, createParams: ContactCreateParams): Promise<Contact> {
+  public async create(customerId: string, connectionId: string, createParams: LeadCreateParams): Promise<Lead> {
     // TODO: We may want to have better guarantees that we update the record in both our DB
     // and the external integration.
     const remoteCreateParams = { ...createParams };
-    if (createParams.accountId) {
-      remoteCreateParams.accountId = await getRemoteId(this.prisma, createParams.accountId, 'account');
-    }
     if (createParams.ownerId) {
       remoteCreateParams.ownerId = await getRemoteId(this.prisma, createParams.ownerId, 'user');
     }
     const remoteClient = (await this.remoteService.getRemoteClient(connectionId)) as CrmRemoteClient;
-    const remoteContact = await remoteClient.createObject('contact', remoteCreateParams);
-    const contactModel = await this.prisma.crmContact.create({
+    const remoteLead = await remoteClient.createObject('lead', remoteCreateParams);
+    const leadModel = await this.prisma.crmLead.create({
       data: {
-        id: uuidv5(remoteContact.remoteId, connectionId),
+        id: uuidv5(remoteLead.remoteId, connectionId),
         customerId,
         connectionId,
-        lastModifiedAt: getLastModifiedAt(remoteContact),
-        ...remoteContact,
-        accountId: createParams.accountId,
+        lastModifiedAt: getLastModifiedAt(remoteLead),
+        ...remoteLead,
         ownerId: createParams.ownerId,
       },
     });
-    return fromContactModel(contactModel);
+    return fromLeadModel(leadModel);
   }
 
-  public async update(customerId: string, connectionId: string, updateParams: ContactUpdateParams): Promise<Contact> {
+  public async update(customerId: string, connectionId: string, updateParams: LeadUpdateParams): Promise<Lead> {
     // TODO: We may want to have better guarantees that we update the record in both our DB
     // and the external integration.
-    const foundContactModel = await this.prisma.crmContact.findUniqueOrThrow({
+    const foundLeadModel = await this.prisma.crmLead.findUniqueOrThrow({
       where: {
         id: updateParams.id,
       },
     });
 
-    if (foundContactModel.customerId !== customerId) {
-      throw new Error('Contact customerId does not match');
+    if (foundLeadModel.customerId !== customerId) {
+      throw new Error('Lead customerId does not match');
     }
 
     const remoteUpdateParams = { ...updateParams };
-    if (updateParams.accountId) {
-      remoteUpdateParams.accountId = await getRemoteId(this.prisma, updateParams.accountId, 'account');
-    }
     if (updateParams.ownerId) {
-      remoteUpdateParams.ownerId = await getRemoteId(this.prisma, updateParams.ownerId, 'user');
+      remoteUpdateParams.ownerId = await await getRemoteId(this.prisma, updateParams.ownerId, 'user');
     }
 
     const remoteClient = (await this.remoteService.getRemoteClient(connectionId)) as CrmRemoteClient;
-    const remoteContact = await remoteClient.updateObject('contact', {
+    const remoteLead = await remoteClient.updateObject('lead', {
       ...remoteUpdateParams,
-      remoteId: foundContactModel.remoteId,
+      remoteId: foundLeadModel.remoteId,
     });
 
-    const contactModel = await this.prisma.crmContact.update({
+    const leadModel = await this.prisma.crmLead.update({
       data: {
-        ...remoteContact,
-        lastModifiedAt: getLastModifiedAt(remoteContact),
-        accountId: updateParams.accountId,
+        ...remoteLead,
+        lastModifiedAt: getLastModifiedAt(remoteLead),
         ownerId: updateParams.ownerId,
       },
       where: {
         id: updateParams.id,
       },
     });
-    return fromContactModel(contactModel);
+    return fromLeadModel(leadModel);
   }
 
-  public async upsertRemoteContacts(
+  public async upsertRemoteLeads(
     connectionId: string,
     customerId: string,
-    remoteContactsReadable: Readable,
+    remoteLeadsReadable: Readable,
     onUpsertBatchCompletion: (offset: number, numRecords: number) => void
   ): Promise<UpsertRemoteCommonModelsResult> {
-    const table = COMMON_MODEL_DB_TABLES['contacts'];
-    const tempTable = 'crm_contacts_temp';
+    const table = COMMON_MODEL_DB_TABLES.crm.leads;
+    const tempTable = 'crm_leads_temp';
     const columnsWithoutId = [
       'remote_id',
       'customer_id',
       'connection_id',
+      'lead_source',
+      'title',
+      'company',
       'first_name',
       'last_name',
       'addresses',
-      'email_addresses',
       'phone_numbers',
-      'lifecycle_stage',
-      'last_activity_at',
+      'email_addresses',
       'remote_created_at',
       'remote_updated_at',
       'remote_was_deleted',
       'remote_deleted_at',
       'detected_or_remote_deleted_at',
       'last_modified_at',
-      '_remote_account_id',
-      'account_id',
+      'converted_date',
+      '_converted_remote_account_id',
+      'converted_account_id',
+      '_converted_remote_contact_id',
+      'converted_contact_id',
       '_remote_owner_id',
       'owner_id',
       'updated_at', // TODO: We should have default for this column in Postgres
@@ -196,11 +191,11 @@ export class ContactService extends CommonModelBaseService {
     return await this.upsertRemoteCommonModels(
       connectionId,
       customerId,
-      remoteContactsReadable,
+      remoteLeadsReadable,
       table,
       tempTable,
       columnsWithoutId,
-      fromRemoteContactToDbContactParams,
+      fromRemoteLeadToDbLeadParams,
       onUpsertBatchCompletion
     );
   }
