+++ packages/core/services/common_models/crm/user_service.ts
@@ -1,10 +1,10 @@
-import { schemaPrefix } from '@supaglue/db';
+import { COMMON_MODEL_DB_TABLES } from '@supaglue/db';
 import { GetInternalParams, ListInternalParams, PaginatedResult, User } from '@supaglue/types';
 import { Readable } from 'stream';
-import { NotFoundError, UnauthorizedError } from '../../errors';
-import { getPaginationParams, getPaginationResult } from '../../lib/pagination';
-import { fromRemoteUserToDbUserParams, fromUserModel } from '../../mappers';
-import { CommonModelBaseService, UpsertRemoteCommonModelsResult } from './base_service';
+import { CommonModelBaseService, UpsertRemoteCommonModelsResult } from '..';
+import { NotFoundError, UnauthorizedError } from '../../../errors';
+import { getPaginationParams, getPaginationResult } from '../../../lib';
+import { fromRemoteUserToDbUserParams, fromUserModel } from '../../../mappers';
 
 export class UserService extends CommonModelBaseService {
   public constructor(...args: ConstructorParameters<typeof CommonModelBaseService>) {
@@ -58,7 +58,7 @@ export class UserService extends CommonModelBaseService {
     remoteUsersReadable: Readable,
     onUpsertBatchCompletion: (offset: number, numRecords: number) => void
   ): Promise<UpsertRemoteCommonModelsResult> {
-    const table = `${schemaPrefix}crm_users`;
+    const table = COMMON_MODEL_DB_TABLES.crm.users;
     const tempTable = 'crm_users_temp';
     const columnsWithoutId = [
       'remote_id',
