+++ packages/core/services/common_models/engagement/index.ts
@@ -1,0 +1,1 @@
+export * from './contact_service';
