+++ packages/db/index.ts
@@ -15,10 +15,15 @@ if (schema) {
 }
 
 export const COMMON_MODEL_DB_TABLES = {
-  contacts: `${schemaPrefix}crm_contacts`,
-  accounts: `${schemaPrefix}crm_accounts`,
-  leads: `${schemaPrefix}crm_leads`,
-  opportunities: `${schemaPrefix}crm_opportunities`,
-  users: `${schemaPrefix}crm_users`,
-  events: `${schemaPrefix}crm_events`,
+  crm: {
+    contacts: `${schemaPrefix}crm_contacts`,
+    accounts: `${schemaPrefix}crm_accounts`,
+    leads: `${schemaPrefix}crm_leads`,
+    opportunities: `${schemaPrefix}crm_opportunities`,
+    users: `${schemaPrefix}crm_users`,
+    events: `${schemaPrefix}crm_events`,
+  },
+  engagement: {
+    contacts: `${schemaPrefix}engagement_contacts`,
+  },
 };
